clean up after PR comments
