+++ ios/Example/Tests/TestFixtures.swift
@@ -13,53 +13,12 @@ class TestFixtures {
     
     public static let rootTrifleCertEncoded3  = Data(base64Encoded: "CAAS6AIwggFkMIIBCqADAgECAgEBMAoGCCqGSM49BAMCMBgxFjAUBgNVBAMMDWlzc3VpbmdFbnRpdHkwHhcNMjMwNTE2MTk1NTIxWhcNMjQwNTE1MTk1NTIxWjAYMRYwFAYDVQQDDA1pc3N1aW5nRW50aXR5MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE8EoJnpQ+1D+IFSq4nkCC/tOaBOxN1pu0TI7qJ0mMnesRiRNGkIrBRCF1OkfSD9vncjHA2XRjeLO6G4RBfUjRu6NFMEMwDwYDVR0TAQH/BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAgQwIAYDVR0OAQH/BBYEFMeBY2S+5Wj+QjLsHW2C4tc42NVSMAoGCCqGSM49BAMCA0gAMEUCIDjJGqIKF4Wci/m97USvUVRQDNXi8oPlLTyNgyjdydbJAiEA2WYjYkUG12UD5iNVWr6w03w84tinujNvvD9ACXALtKI=")
 
-    public static let deviceCertEncoded = Data(base64Encoded: """
-MIIBFDCBvKADAgECAgYBhtoH23kwCgYIKoZIzj0EAwIwEzERMA8GA1\
-UEAwwIY2FzaC5hcHAwHhcNMjMwMzEzMDgxMzEzWhcNMjMwOTA5MDgx\
-MzEzWjARMQ8wDQYDVQQDDAZlbnRpdHkwWTATBgcqhkjOPQIBBggqhk\
-jOPQMBBwNCAASGmNOZtRsOHr2XJVjL1jxRpeumY/jAF3rBDXhhbP49\
-5VYvMYkHfWjeNBzJ1YIqXLQf4BHdQQzDjSvT+abzpdjEMAoGCCqGSM\
-49BAMCA0cAMEQCIBXBwAm0Qn7S4vKRIHYLXDxhUQTqMuReoWBPfiOX\
-fREnAiByNBjSwVxp980/Nl17UvkBUNhTvp55CNuwMmlnporNNA==
-""")
-    
     public static let  deviceTrifleCertEncoded = Data(base64Encoded: "CAASmAIwggEUMIG8oAMCAQICBgGG2gfbeTAKBggqhkjOPQQDAjATMREwDwYDVQQDDAhjYXNoLmFwcDAeFw0yMzAzMTMwODEzMTNaFw0yMzA5MDkwODEzMTNaMBExDzANBgNVBAMMBmVudGl0eTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABIaY05m1Gw4evZclWMvWPFGl66Zj+MAXesENeGFs/j3lVi8xiQd9aN40HMnVgipctB/gEd1BDMONK9P5pvOl2MQwCgYIKoZIzj0EAwIDRwAwRAIgFcHACbRCftLi8pEgdgtcPGFRBOoy5F6hYE9+I5d9EScCIHI0GNLBXGn3zT82XXtS+QFQ2FO+nnkI27AyaWemis00")
                                                         
-    public static let deviceCertEncoded2 = Data(base64Encoded: """
-MIIBGTCBwaADAgECAgYBht0eNMMwCgYIKoZIzj0EAwIwGDEWMBQGA1\
-UEAwwNaXNzdWluZ0VudGl0eTAeFw0yMzAzMTMyMjM2MjlaFw0yMzA5\
-MDkyMjM2MjlaMBExDzANBgNVBAMMBmVudGl0eTBZMBMGByqGSM49Ag\
-EGCCqGSM49AwEHA0IABICead8cmQi2cyHHTx316w9Q64L11U86PV3R\
-K1IDsm/xiDoa5sbShjFPm0nhd+AFoTPtsXL6SJ/bt+sndXQL5gcwCg\
-YIKoZIzj0EAwIDRwAwRAIgBQLsaQZpa93v33J/kSIxcl2UtBPCyYYD\
-KahIGLy7xM4CIGeiGFjglmmaiqFf30esHdL4yR0/rbkVm4h6z9O+Rjfp
-""")
-    
     public static let deviceTrifleCertEncoded2 = Data(base64URLEncoded: "CAASnQIwggEZMIHBoAMCAQICBgGG3R40wzAKBggqhkjOPQQDAjAYMRYwFAYDVQQDDA1pc3N1aW5nRW50aXR5MB4XDTIzMDMxMzIyMzYyOVoXDTIzMDkwOTIyMzYyOVowETEPMA0GA1UEAwwGZW50aXR5MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEgJ5p3xyZCLZzIcdPHfXrD1DrgvXVTzo9XdErUgOyb_GIOhrmxtKGMU-bSeF34AWhM-2xcvpIn9u36yd1dAvmBzAKBggqhkjOPQQDAgNHADBEAiAFAuxpBmlr3e_fcn-RIjFyXZS0E8LJhgMpqEgYvLvEzgIgZ6IYWOCWaZqKoV_fR6wd0vjJHT-tuRWbiHrP075GN-k")
         
-    public static let otherRootCertEncoded = Data(base64Encoded: """
-MIHcMIGPoAMCAQICAQEwBQYDK2VwMBgxFjAUBgNVBAMMDWlzc3VpbmdFbnR\
-pdHkwHhcNMjMwMzEzMDYxMzI3WhcNMjMwMzE0MDYxMzI3WjAYMRYwFAYDVQ\
-QDDA1pc3N1aW5nRW50aXR5MCowBQYDK2VwAyEAm+Ac7932SHDPQLYd3p3gm\
-grcArUWrBqhPEC+q/QI3lEwBQYDK2VwA0EAJrfzN7qA3VqwazsT8yMIYMvY\
-Rz2iDA1898Yx5ELtlQcl7QUGXUmadwzW7rpxQB5wIk46tPTEJCFmUIYwrCB\
-4BQ==
-""")
-    
     public static let otherRootTrifleCertEncoded = Data(base64Encoded: "CAAS3wEwgdwwgY+gAwIBAgIBATAFBgMrZXAwGDEWMBQGA1UEAwwNaXNzdWluZ0VudGl0eTAeFw0yMzAzMTMwNjEzMjdaFw0yMzAzMTQwNjEzMjdaMBgxFjAUBgNVBAMMDWlzc3VpbmdFbnRpdHkwKjAFBgMrZXADIQCb4Bzv3fZIcM9Ath3eneCaCtwCtRasGqE8QL6r9AjeUTAFBgMrZXADQQAmt/M3uoDdWrBrOxPzIwhgy9hHPaIMDXz3xjHkQu2VByXtBQZdSZp3DNbuunFAHnAiTjq09MQkIWZQhjCsIHgF")
 
-    public static let rootCertEncoded = Data(base64Encoded: """
-MIIBZTCCAQqgAwIBAgIBATAKBggqhkjOPQQDAjAYMRYwFAYDVQQDDA1pc3N\
-1aW5nRW50aXR5MB4XDTIzMDMxMzIyMzUzMloXDTIzMDkwOTIyMzUzMlowGD\
-EWMBQGA1UEAwwNaXNzdWluZ0VudGl0eTBZMBMGByqGSM49AgEGCCqGSM49A\
-wEHA0IABICead8cmQi2cyHHTx316w9Q64L11U86PV3RK1IDsm/xiDoa5sbS\
-hjFPm0nhd+AFoTPtsXL6SJ/bt+sndXQL5gejRTBDMA8GA1UdEwEB/wQFMAM\
-BAf8wDgYDVR0PAQH/BAQDAgIEMCAGA1UdDgEB/wQWBBQ/80Y00UVTlI6kiA\
-ZZ46kcrJ9a2jAKBggqhkjOPQQDAgNJADBGAiEAvffuwvImKNaolqnEr4ENB\
-6LXEFdV0YVK3Ic3Mi+hqJ8CIQC7CiLwyvH1cChUReanIGeYiQp27LJ99M/q\
-WLq6hmtSmQ==
-""")
-    
     public static let rootTrifleCertEncoded = Data(base64Encoded: "CAAS6QIwggFlMIIBCqADAgECAgEBMAoGCCqGSM49BAMCMBgxFjAUBgNVBAMMDWlzc3VpbmdFbnRpdHkwHhcNMjMwMzEzMjIzNTMyWhcNMjMwOTA5MjIzNTMyWjAYMRYwFAYDVQQDDA1pc3N1aW5nRW50aXR5MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEgJ5p3xyZCLZzIcdPHfXrD1DrgvXVTzo9XdErUgOyb/GIOhrmxtKGMU+bSeF34AWhM+2xcvpIn9u36yd1dAvmB6NFMEMwDwYDVR0TAQH/BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAgQwIAYDVR0OAQH/BBYEFD/zRjTRRVOUjqSIBlnjqRysn1raMAoGCCqGSM49BAMCA0kAMEYCIQC99+7C8iYo1qiWqcSvgQ0HotcQV1XRhUrchzcyL6GonwIhALsKIvDK8fVwKFRF5qcgZ5iJCnbssn30z+pYurqGa1KZ")
     
     // https://www.sslshopper.com/certificate-decoder.html
